+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `Innmind\Http\Headers::filter()`
+- `Innmind\Http\Message\Method::safe()` and `::idempotent()`
 
 ## 6.1.0 - 2023-02-17